// src/models/user/attributes/UserAttributeCategory.ts
export type UserAttributeCategory =
  | 'Figur'
  | 'Raucher'
  | 'Alkohol'
  | 'Wohnen'
  | 'Beziehung'
  | 'Orientierung'
  | 'Kinder'
  | 'Haarfarbe'
  | string;

export enum UserAttributeCategories {
  SHAPE = 'Figur',
  SMOKER = 'Raucher',
  ALCOHOL = 'Alkohol',
  LIVING = 'Wohnen',
  RELATION = 'Beziehung',
  ORIENTATION = 'Orientierung',
  CHILDREN = 'Kinder',
  HAIRCOLOR = 'Haarfarbe',
}